requirements.txt
astroid==2.8.2
cachetools==4.1.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
click-log==0.3.2
cloudinary==1.22.0
cloudinary-cli==1.1.1
Flask-SQLAlchemy==2.4.3
grpcio==1.31.0
idna==2.9
isort==4.3.21
itsdangerous==1.1.
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==7.2.0
protobuf==3.13.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pylint==2.5.3
python-dateutil==2.8.1
python-magic==0.4.18
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.17
sqlalchemy-media==0.17.4
text-unidecode==1.3
toml==0.10.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1